refactor: separated createWaveSurfer
