+++ .gitignore
@@ -12,3 +12,4 @@ vite.config.js.timestamp-*
 vite.config.ts.timestamp-*
 /.cargo
 /.xwin-cache
+/.bfg-report
