+++ src/routes/+page.svelte
@@ -1,14 +1,8 @@
 <script lang="ts">
-	import { readBinaryFile } from '@tauri-apps/api/fs';
-	import WaveSurfer from 'wavesurfer.js';
-	import RegionsPlugin, { type Region } from 'wavesurfer.js/dist/plugins/regions.js';
-	import TimelinePlugin from 'wavesurfer.js/dist/plugins/timeline';
-	import Hover from 'wavesurfer.js/dist/plugins/hover';
+	import type WaveSurfer from 'wavesurfer.js';
+	import type RegionsPlugin from 'wavesurfer.js/dist/plugins/regions.js';
 	import {
 		appProcessStatusReadableStore,
-		appProcessStatusWritableStore,
-		regionStore,
-		playerStore,
 		currentTimeStore,
 		durationStore,
 		currentFileUrlStore,
@@ -17,22 +11,13 @@
 	import { onDestroy } from 'svelte';
 	import Footer from '../components/Footer.svelte';
 	import AudioCutterControls from '../components/AudioCutterControls.svelte';
-	import {
-		cHoverOnWaveBar,
-		cBackground,
-		cUnplayedSection,
-		cPlayedSection,
-		cSelectedRegion,
-		cText
-	} from './styles';
+
 	import { loadAudio } from './loadAudio';
 	import { get } from 'svelte/store';
-	let loop = true;
-	let minPxPerSecBindValue = 100;
+	import { createWaveSurfer } from './createWaveSurfer';
 	let ws: WaveSurfer;
 	let wsRegions: RegionsPlugin;
-
-	let binaryAudio;
+	let waveformDiv: HTMLElement;
 
 	function formatTime(seconds: number) {
 		if (seconds === undefined) return '0:00';
@@ -40,147 +25,6 @@
 		let secs = Math.floor(seconds % 60);
 
 		return `${minutes}:${secs > 10 ? secs : `0${secs}`}`;
-	}
-
-	async function createWaveSurfer() {
-		/* read data into a Uint8Array */
-		console.log('Begin blob reading');
-		appProcessStatusWritableStore.set('Reading file...');
-
-		binaryAudio = await readBinaryFile(get(currentFileUrlStore));
-		appProcessStatusWritableStore.set('Reading is done.');
-
-		console.log('Blob reading is done');
-
-		if (ws !== undefined) {
-			ws.destroy();
-			regionStore.update((value) => {
-				value.start = 0;
-				value.end = 0;
-				return value;
-			});
-		}
-
-		const bottomTimeline = TimelinePlugin.create({
-			height: 28,
-			timeInterval: 1,
-			primaryLabelInterval: 1,
-			style: {
-				fontSize: '12px',
-				color: cText
-			}
-		});
-		const hover = Hover.create({
-			lineColor: cHoverOnWaveBar,
-			lineWidth: 2,
-			labelBackground: cBackground,
-			labelColor: cText,
-			labelSize: '11px'
-		});
-
-		ws = WaveSurfer.create({
-			container: '#waveform',
-			waveColor: cUnplayedSection,
-			progressColor: cPlayedSection,
-			plugins: [bottomTimeline, hover],
-			minPxPerSec: minPxPerSecBindValue
-		});
-		console.log('wsinit', ws);
-		ws.getMediaElement().preload = 'auto';
-
-		ws.getMediaElement().onloadedmetadata = () => {
-			durationStore.set(String(ws.getMediaElement().duration));
-		};
-
-		setTimeout(() => appProcessStatusWritableStore.set('😉'), 2000);
-		await ws.loadBlob(new Blob([binaryAudio.buffer]));
-
-		/** When the audio starts playing */
-		ws.on('play', () => {
-			appProcessStatusWritableStore.set('Playing.');
-			playerStore.update((store) => ({ ...store, play: true }));
-			console.log('Play');
-		});
-
-		/** When the audio pauses */
-		ws.on('pause', () => {
-			appProcessStatusWritableStore.set('Paused.');
-			playerStore.update((store) => ({ ...store, play: false }));
-			console.log('Pause');
-		});
-
-		ws.on('finish', () => {
-			console.log('Finish');
-			playerStore.update((store) => ({ ...store, play: false }));
-			appProcessStatusWritableStore.set('Stopped.');
-		});
-
-		/** On audio position change, fires continuously during playback */
-		ws.on('timeupdate', (currentTime) => {
-			currentTimeStore.set(String(currentTime));
-		});
-
-		/** When the user interacts with the waveform (i.g. clicks or drags on it) */
-		ws.on('interaction', (newTime) => {
-			activeRegion = null;
-
-			console.log('Interaction', newTime + 's');
-		});
-
-		/** When the zoom level changes */
-		ws.on('zoom', (minPxPerSecBindValue) => {
-			console.log('Zoom', minPxPerSecBindValue + 'px/s');
-		});
-
-		wsRegions = ws.registerPlugin(RegionsPlugin.create());
-
-		wsRegions.on('region-created', (region) => {
-			wsRegions.getRegions().forEach((r) => {
-				if (r.id !== region.id) {
-					r.remove();
-				}
-			});
-			regionStore.update((value) => {
-				value.start = region.start;
-				value.end = region.end;
-				return value;
-			});
-		});
-		wsRegions.on('region-updated', (region) => {
-			console.log('Updated region', region);
-			regionStore.update((value) => {
-				value.start = region.start;
-				value.end = region.end;
-				return value;
-			});
-		});
-
-		wsRegions.enableDragSelection({
-			color: cSelectedRegion
-		});
-
-		let activeRegion: Region | null = null;
-		wsRegions.on('region-in', (region) => {
-			activeRegion = region;
-		});
-
-		wsRegions.on('region-out', (region) => {
-			if (activeRegion === region) {
-				if (loop) {
-					region.play();
-				} else {
-					activeRegion = null;
-				}
-			}
-		});
-
-		wsRegions.on('region-clicked', (region: Region, e: Event) => {
-			e.stopPropagation();
-			activeRegion = region;
-			region.play();
-		});
-
-		isAudioSelectedStore.set(true);
 	}
 
 	$: currentTime = '';
@@ -218,7 +62,8 @@
 			</div>
 		{/if}
 		<div>
-			<div class={$isAudioSelectedStore ? '' : 'hidden'} id="waveform" />
+			<div id="waveform" class={$isAudioSelectedStore ? '' : 'hidden'} />
+
 			<div class={currentTime || duration ? 'flex' : 'hidden'} id="time-and-duration">
 				<div id="time">{formatTime(Number(currentTime))}</div>
 
@@ -226,7 +71,7 @@
 			</div>
 		</div>
 		{#if $isAudioSelectedStore === true}
-			<AudioCutterControls {ws} {wsRegions} bind:loop />
+			<AudioCutterControls {ws} {wsRegions} />
 		{/if}
 		{#if $appProcessStatusReadableStore !== 'Reading file...'}
 			<div class="flex items-center justify-center">
@@ -238,7 +83,10 @@
 						if (!shouldLoadAudio) {
 							return;
 						}
-						await createWaveSurfer();
+						let wsAndWsRegions = await createWaveSurfer(ws, wsRegions);
+						ws = wsAndWsRegions.ws;
+						wsRegions = wsAndWsRegions.wsRegions;
+						isAudioSelectedStore.set(true);
 					}}
 				>
 					{#if $isAudioSelectedStore === null}
