+++ src/routes/styles.css
@@ -20,6 +20,7 @@ html {
 	tab-size: 4;
 	font-family: 'Inria Sans', sans-serif;
 	font-size: 1em;
+	color: var(--text);
 }
 body {
 	margin: 0;
